+++ plotnine/geoms/geom_vline.py
@@ -9,7 +9,7 @@ import pandas as pd
 from ..doctools import document
 from ..exceptions import PlotnineWarning
 from ..mapping import aes
-from ..utils import SIZE_FACTOR, order_as_data_mapping
+from ..utils import SIZE_FACTOR, order_as_data_mapping, to_rgba
 from .geom import geom
 from .geom_segment import geom_segment
 
@@ -121,13 +121,13 @@ class geom_vline(geom):
         x = [0.5 * da.width] * 2
         y = [0, da.height]
         data["size"] *= SIZE_FACTOR
+        color = to_rgba(data["color"], data["alpha"])
         key = Line2D(
             x,
             y,
-            alpha=data["alpha"],
             linestyle=data["linetype"],
             linewidth=data["size"],
-            color=data["color"],
+            color=color,
             solid_capstyle="butt",
             antialiased=False,
         )